DyGLib/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
appdirs=1.4.4=pyh9f0ad1d_0
asttokens=2.4.1=pyhd8ed1ab_0
atk-1.0=2.36.0=ha1a6a79_0
blas=1.0=mkl
boost-cpp=1.82.0=hdb19cb5_2
brotli=1.1.0=hd590300_1
brotli-bin=1.1.0=hd590300_1
brotli-python=1.1.0=py311hb755f60_1
bzip2=1.0.8=hd590300_5
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.3.11=h06a4308_0
cairo=1.16.0=hb05425b_5
certifi=2024.2.2=py311h06a4308_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=unix_pyh707e725_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.2.0=py311h9547e67_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.3.101=0
cuda-runtime=12.1.0=0
cycler=0.12.1=pyhd8ed1ab_0
dbus=1.13.18=hb2f20db_0
debugpy=1.8.0=py311hb755f60_1
decorator=5.1.1=pyhd8ed1ab_0
docker-pycreds=0.4.0=py_0
einops=0.7.0=pypi_0
exceptiongroup=1.2.0=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
expat=2.5.0=hcb278e6_1
fairscale=0.4.13=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=2.001=hcb22688_0
font-ttf-source-code-pro=2.030=hd3eb1b0_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.14.2=h14ed4e7_0
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
fonttools=4.47.0=py311h459d7ec_0
freetype=2.12.1=h267a509_2
fribidi=1.0.10=h7b6447c_0
fsspec=2023.12.2=pypi_0
gdk-pixbuf=2.42.10=h5eee18b_0
gettext=0.21.1=h27087fc_0
giflib=5.2.1=h0b41bf4_3
gitdb=4.0.11=pyhd8ed1ab_0
gitpython=3.1.40=pyhd8ed1ab_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.3.0=h59595ed_0
gmpy2=2.1.2=py311h6a5fa03_1
gnutls=3.6.13=h85f3911_1
gobject-introspection=1.78.1=py311hd56c82f_0
graphite2=1.3.14=h295c915_1
graphviz=2.50.0=h1b29801_1
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
gtk2=2.24.33=h73c1081_2
gts=0.7.6=hb67d8dd_3
harfbuzz=4.3.0=hf52aaf7_2
icecream=2.1.3=pyhd8ed1ab_0
icu=73.2=h59595ed_0
idna=3.6=pyhd8ed1ab_0
importlib-metadata=7.0.1=pyha770c72_0
importlib_metadata=7.0.1=hd8ed1ab_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.28.0=pyhd33586a_0
ipython=8.19.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.3.2=pyhd8ed1ab_0
jpeg=9e=h166bdaf_2
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.6.0=py311h38be061_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.5=py311h9547e67_1
krb5=1.20.1=h81ceb04_0
labml=0.4.168=pypi_0
labml-helpers=0.4.89=pypi_0
lame=3.100=h166bdaf_1003
lcms2=2.15=hfd0df8a_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=4.0.0=h27087fc_0
libabseil=20230802.1=cxx17_h59595ed_0
libauc=1.3.0=pypi_0
libblas=3.9.0=1_h86c2bf4_netlib
libboost=1.82.0=h109eef0_2
libbrotlicommon=1.1.0=hd590300_1
libbrotlidec=1.1.0=hd590300_1
libbrotlienc=1.1.0=hd590300_1
libcblas=3.9.0=5_h92ddd45_netlib
libclang=14.0.6=default_h7634d5b_1
libclang13=14.0.6=default_h9986a30_1
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.8.1.2=0
libcups=2.3.3=h36d4200_3
libcurand=10.3.4.101=0
libcurl=8.5.0=h251f7ec_0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h0b41bf4_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h7f8727e_1
libexpat=2.5.0=hcb278e6_1
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h807b86a_3
libgd=2.3.3=h695aa2c_1
libgfortran-ng=13.2.0=h69a702a_3
libgfortran5=13.2.0=ha4646dd_3
libglib=2.78.4=hdc74915_0
libgomp=13.2.0=h807b86a_3
libiconv=1.17=hd590300_2
libjpeg-turbo=2.0.0=h9bf148f_0
liblapack=3.9.0=5_h92ddd45_netlib
libllvm14=14.0.6=hcd5def8_4
libnghttp2=1.57.0=h2d74bed_0
libnpp=12.0.2.50=0
libnsl=2.0.1=hd590300_0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h753d276_0
libpq=12.15=hdbd6064_1
libprotobuf=4.24.4=hf27288f_0
librsvg=2.54.4=h36cc946_3
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.44.2=h2797004_0
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=13.2.0=h7e041cc_3
libtiff=4.5.0=h6adf6a1_2
libtool=2.4.6=h6a678d5_1009
libuuid=2.38.1=h0b41bf4_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=hd590300_0
libxcb=1.15=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hf1b16e4_1
libzlib=1.2.13=hd590300_5
littleutils=0.2.2=py_0
llvm-openmp=15.0.7=h0cdce71_0
markupsafe=2.1.3=py311h459d7ec_1
matplotlib=3.8.2=py311h38be061_0
matplotlib-base=3.8.2=py311h54ef318_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2023.1.0=h213fc3f_46344
mpc=1.3.1=hfe3b2da_0
mpfr=4.2.1=h9458935_0
mpmath=1.3.0=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
mysql=5.7.20=hf484d3e_1001
ncurses=6.4=h59595ed_2
nest-asyncio=1.5.8=pyhd8ed1ab_0
nettle=3.6=he412f7d_0
networkx=3.2.1=pyhd8ed1ab_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numpy=1.26.2=py311h64a7726_0
ogb=1.3.6=pyhd8ed1ab_0
opencv-python=4.8.1.78=pypi_0
openh264=2.1.1=h780b84a_0
openjpeg=2.5.0=hfec8fc6_2
openssl=3.2.1=hd590300_0
outdated=0.2.2=pyhd8ed1ab_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.4=py311h320fe9a_0
pango=1.50.7=h05da053_0
parso=0.8.3=pyhd8ed1ab_0
pathtools=0.1.2=py_1
pcre2=10.42=hcad00b1_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.0.1=py311ha6cbd5a_0
pip=23.3.2=pyhd8ed1ab_0
pixman=0.40.0=h7f8727e_1
platformdirs=4.1.0=pyhd8ed1ab_0
plotly=5.18.0=pyhd8ed1ab_0
ply=3.11=py_1
poppler=22.12.0=h9614445_3
poppler-data=0.4.11=h06a4308_1
prompt-toolkit=3.0.42=pyha770c72_0
protobuf=4.24.4=py311h46cbc50_0
psutil=5.9.7=py311h459d7ec_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pybind11=2.11.1=py311h9547e67_2
pybind11-global=2.11.1=py311h9547e67_2
pyg=2.4.0=py311_torch_2.1.0_cu121
pyg-lib=0.3.1+pt21cu121=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pyparsing=3.1.1=pyhd8ed1ab_0
pyqt=5.15.10=py311h6a678d5_0
pyqt5-sip=12.13.0=py311h5eee18b_0
pysocks=1.7.1=pyha2e5f31_6
python=3.11.7=hab00c5b_1_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.11=4_cp311
pytorch=2.1.2=py3.11_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pyhd8ed1ab_0
pyyaml=6.0.1=py311h459d7ec_1
pyzmq=25.1.2=py311h34ded2d_0
qt-main=5.15.2=h53bd1ea_10
readline=8.2=h8228510_1
requests=2.31.0=pyhd8ed1ab_0
scikit-learn=1.3.2=py311hc009520_2
scipy=1.11.4=py311h64a7726_0
sentry-sdk=1.39.1=pyhd8ed1ab_0
setproctitle=1.3.3=py311h459d7ec_0
setuptools=68.2.2=pyhd8ed1ab_0
sip=6.7.12=py311hb755f60_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
sqlite=3.44.2=h2c6b66d_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypyh9d50eac_103
tbb=2021.8.0=hdb19cb5_0
tenacity=8.2.3=pyhd8ed1ab_0
threadpoolctl=3.2.0=pyha21a80b_0
tk=8.6.13=noxft_h4845f30_101
tomli=2.0.1=pyhd8ed1ab_0
torch-cluster=1.6.3+pt21cu121=pypi_0
torch-scatter=2.1.2+pt21cu121=pypi_0
torch-sparse=0.6.18+pt21cu121=pypi_0
torch-spline-conv=1.2.2+pt21cu121=pypi_0
torchaudio=2.1.2=py311_cu121
torchdata=0.7.1=pypi_0
torchtext=0.16.2=pypi_0
torchtriton=2.1.0=py311
torchvision=0.16.2=py311_cu121
tornado=6.3.3=py311h459d7ec_1
tqdm=4.66.1=pyhd8ed1ab_0
traitlets=5.14.0=pyhd8ed1ab_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023d=h0c530f3_0
urllib3=2.1.0=pyhd8ed1ab_0
wandb=0.16.3=pyhd8ed1ab_0
wcwidth=0.2.12=pyhd8ed1ab_0
wheel=0.42.0=pyhd8ed1ab_0
xkeyboard-config=2.40=hd590300_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libx11=1.8.7=h8ee46fc_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xproto=7.0.31=h7f98852_1007
xz=5.4.6=h5eee18b_0
yaml=0.2.5=h7f98852_2
zeromq=4.3.5=h59595ed_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=hd590300_5
zstd=1.5.5=hfc55251_0
